Skip auto-lowmem logic for large files
In Tarsnap 1.0.15 (November 2008) code was added to switch into a
quasi-lowmem mode if it finds that a lot of memory is being used to
cache small files.

Adjust that behaviour to only kick in for files which are < 10 MB
in size; larger files will be fully cached (unless --lowmem or
--verylowmem are specified explicitly) so they don't need to be
re-read on every run.

Reported by:	Tim Bishop